+++ Commands/Graph/RemoveMicrosoft365Group.cs
@@ -8,28 +8,29 @@ using System.Management.Automation;
 
 namespace SharePointPnP.PowerShell.Commands.Graph
 {
-    [Cmdlet(VerbsCommon.Remove, "PnPUnifiedGroup")]
-    [CmdletHelp("Removes one Microsoft 365 Group (aka Unified Group)",
+    [Cmdlet(VerbsCommon.Remove, "PnPMicrosoft365Group")]
+    [Alias("Remove-PnPUnifiedGroup")]
+    [CmdletHelp("Removes one Microsoft 365 Group",
         Category = CmdletHelpCategory.Graph,
         OutputTypeLink = "https://docs.microsoft.com/graph/api/group-delete",
         SupportedPlatform = CmdletSupportedPlatform.Online)]
     [CmdletExample(
-       Code = "PS:> Remove-PnPUnifiedGroup -Identity $groupId",
+       Code = "PS:> Remove-PnPMicrosoft365Group -Identity $groupId",
        Remarks = "Removes an Microsoft 365 Group based on its ID",
        SortOrder = 1)]
     [CmdletExample(
-       Code = "PS:> Remove-PnPUnifiedGroup -Identity $group",
+       Code = "PS:> Remove-PnPMicrosoft365Group -Identity $group",
        Remarks = "Removes the provided Microsoft 365 Group",
        SortOrder = 2)]
     [CmdletExample(
-       Code = "PS:> Get-PnPUnifiedGroup | ? Visibility -eq \"Public\" | Remove-PnPUnifiedGroup",
+       Code = "PS:> Get-PnPMicrosoft365Group | ? Visibility -eq \"Public\" | Remove-PnPMicrosoft365Group",
        Remarks = "Removes all the public Microsoft 365 Groups",
        SortOrder = 3)]
     [CmdletMicrosoftGraphApiPermission(MicrosoftGraphApiPermission.Group_ReadWrite_All)]
-    public class RemoveUnifiedGroup : PnPGraphCmdlet
+    public class RemoveMicrosoft365Group : PnPGraphCmdlet
     {
         [Parameter(Mandatory = true, ValueFromPipeline = true, HelpMessage = "The Identity of the Microsoft 365 Group")]
-        public UnifiedGroupPipeBind Identity;
+        public Microsoft365GroupPipeBind Identity;
 
         protected override void ExecuteCmdlet()
         {